> -Write methods still has some memory issues probably with the > record_handler callback function. Any ideas? Corrected the segmentation fault in record_handler callback function. "packedsamples" need to be a pointer. I also updated the libmseed.so (compiled with the -g debugging flag). Debugging can be done like this::
    ddd /usr/bin/python
    (gdb) run myfile.py
    "File"->"Open Source"->"mysourcefile"
    ...

> -Added test_visual.py for a visual sanity test of the Mini-SEED read
> method. (Only for testing purposes - Needs gnuplot to work)
matplotlib makes plotting much easier and does not rely on gnuplot, added
some lines to test_visual

+++ obspy.mseed/trunk/obspy/mseed/tests/test_visual.py
@@ -1,5 +1,6 @@
 """
-Displays the waveform data from a given Mini-SEED file using gnuplot
+Displays the waveform data from a given Mini-SEED file using gnuplot or
+matplotlib
 
 Only works with Linux and installed gnuplot
 If it does not work make sure gnuplot is in $PATH
@@ -37,27 +38,36 @@ print 'Reading data...'
 mseed=libmseed()
 header, data, numtraces =mseed.read_ms(filename)
 
-#Writes the file containing the data
-print 'Processing data...'
-datafile=file('dat.tmp','w')
-for _i in range(data.__len__()):
-    datafile.write(str(_i)+" "+str(data[_i])+"\n")
-datafile.close()
 
-#Writes the gnuplot configuration file
-print 'Configuring gnuplot...'
-gnufile=file('gnuplotconf.tmp','w')
-gnufile.write('set terminal png\n')
-gnufile.write('set output "'+output+'"\n')
-gnufile.write('set term png size '+str(xres)+','+str(yres)+'\n')
-gnufile.write('set nokey\n')
-gnufile.write('set title "File: '+filename+'"\n')
-gnufile.write('plot "dat.tmp" with lines \n')
-gnufile.close()
+if True:
+    # plotting with matplotlib
+    from pylab import *
+    plot(array(data))
+    savefig(output)
+else:
+    #Writes the file containing the data
+    print 'Processing data...'
+    datafile=file('dat.tmp','w')
+    for _i in range(data.__len__()):
+        datafile.write(str(_i)+" "+str(data[_i])+"\n")
+    datafile.close()
 
-print 'Creating graph...'
-os.system('gnuplot gnuplotconf.tmp')
-print 'Cleaning up temporary files...'
-os.remove('dat.tmp')
-os.remove('gnuplotconf.tmp')
+    #Writes the gnuplot configuration file
+    print 'Configuring gnuplot...'
+    gnufile=file('gnuplotconf.tmp','w')
+    gnufile.write('set terminal png\n')
+    gnufile.write('set output "'+output+'"\n')
+    gnufile.write('set term png size '+str(xres)+','+str(yres)+'\n')
+    gnufile.write('set nokey\n')
+    gnufile.write('set title "File: '+filename+'"\n')
+    gnufile.write('plot "dat.tmp" with lines \n')
+    gnufile.close()
+    
+    print 'Creating graph...'
+    os.system('gnuplot gnuplotconf.tmp')
+    print 'Cleaning up temporary files...'
+    os.remove('dat.tmp')
+    os.remove('gnuplotconf.tmp')
+
+
 print 'Done: Plotted '+filename+' to '+output+'.'
